Move utils type from api.tsx to gridHistory.tsx and delete unncessary conditional statements in HistoricalGrid.tsx
